Remove 5.x obsolete options checks
Since we are moving to 7.0 its a good time to update the code and remove
deprecated code and warning for removed options.

Remove any 5.x options
Add tests for obsolete Filebeat 6.x options
