+++ filebeat/tests/system/test_base.py
@@ -26,26 +26,6 @@ class Test(BaseTest):
         assert "@timestamp" in output
         assert "input.type" in output
 
-    def test_invalid_config_with_removed_settings(self):
-        """
-        Checks if filebeat fails to load if removed settings have been used:
-        """
-        self.render_config_template()
-
-        exit_code = self.run_beat(extra_args=[
-            "-E", "filebeat.spool_size=2048",
-            "-E", "filebeat.publish_async=true",
-            "-E", "filebeat.idle_timeout=1s",
-        ])
-
-        assert exit_code == 1
-        assert self.log_contains("setting 'filebeat.spool_size'"
-                                 " has been removed")
-        assert self.log_contains("setting 'filebeat.publish_async'"
-                                 " has been removed")
-        assert self.log_contains("setting 'filebeat.idle_timeout'"
-                                 " has been removed")
-
     @unittest.skipUnless(INTEGRATION_TESTS, "integration test")
     def test_template(self):
         """
